requirements.txt
black==21.9b0
click==8.0.1
mypy==0.910
mypy-extensions==0.4.3
numpy==1.21.2
pandas==1.3.3
pathspec==0.9.0
platformdirs==2.3.0
pyaes==1.6.1
pyasn1==0.4.8
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
rsa==4.7.2
six==1.16.0
Telethon==1.23.0
toml==0.10.2
tomli==1.2.1
types-PyYAML==5.4.10
typing-extensions==3.10.0.2
